pyproject.toml
[tool.poetry]
name = "kdt-final-django"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
license = "\"MIT\""
readme = "README.md"
packages = [{include = "kdt_final_django"}]

[tool.poetry.dependencies]
python = "~3.9"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
requests = "^2.28.2"
django-cloudflare-images = "^0.4.1"
transformers = "^4.26.0"
scipy = "^1.10.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
